Fix mutter's IdleMonitor cast

+++ base/platform/linux/base_last_input_linux.cpp
@@ -154,7 +154,7 @@ std::optional<crl::time> MutterDBusLastUserInputTime() {
 			{},
 			"org.gnome.Mutter.IdleMonitor");
 
-		const auto value = GlibVariantCast<uint>(reply.get_child(0));
+		const auto value = GlibVariantCast<guint64>(reply.get_child(0));
 		return (crl::now() - static_cast<crl::time>(value));
 	} catch (const Glib::Error &e) {
 		static const auto NotSupportedErrors = {
